add note for using non-LTS release

+++ install/install_docker.sh
@@ -1,5 +1,10 @@
 #!/bin/bash
 
+if [ $EUID != 0 ]; then
+    sudo "$0" "$@"
+    exit $?
+fi
+
 # remove any installed versions
 sudo apt-get remove docker docker-engine docker.io containerd runc
 
@@ -17,10 +22,12 @@ curl -fsSL https://download.docker.com/linux/ubuntu/gpg | sudo apt-key add -
 sudo apt-key fingerprint 0EBFCD88
 
 # add repository
-sudo add-apt-repository \
+sudo add-apt-repository --yes \
    "deb [arch=amd64] https://download.docker.com/linux/ubuntu \
    $(lsb_release -cs) \
    stable"
+  # note : if using a non-LTS release, you may need to replace
+  # $(lsb_release -cs) with the specfic LTS name, i.e. xenial for 20.04
 
 sudo apt-get update
 sudo apt-get install --yes docker-ce docker-ce-cli containerd.io
